move stream agg into new file
